+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+dev
+===
+
+* [Feature] python 3.8 support
+
+
 0.9.26
 ======